+++ Google/google_module/metrics.py
@@ -1,4 +1,4 @@
-from prometheus_client import Counter, Histogram
+from prometheus_client import Counter, Histogram, Gauge
 
 # Declare google prometheus metrics
 prom_namespace_google = "symphony_module_google"
@@ -20,7 +20,7 @@ OUTCOMING_EVENTS = Counter(
     labelnames=["intake_key"],
 )
 
-EVENTS_LAG = Histogram(
+EVENTS_LAG = Gauge(
     name="events_lags",
     documentation="The delay, in seconds, from the date of the last event",
     namespace=prom_namespace,