+++ Proofpoint/proofpoint_modules/trigger_pod_events.py
@@ -165,7 +165,7 @@ class EventsForwarder(Thread):
             # compute the lag
             now = datetime.now(timezone.utc)
             current_lag = now - most_recent_date_seen
-            EVENTS_LAG.labels(intake_key=self.configuration.intake_key).observe(int(current_lag.total_seconds()))
+            EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(int(current_lag.total_seconds()))
 
         return events