+++ Azure/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 2024-02-13 - 2.4.2
+
+### Changed
+
+- Change event_lags metrics from Histogram to Gauge
+
 ## 2024-01-23 - 2.2.4
 
 ### Changed